+++ src/test/java/com/ftn/cdss/monitoring/LowOxygenBloodLevelTest.java
@@ -31,11 +31,11 @@ import static org.junit.Assert.assertThat;
 
 public class LowOxygenBloodLevelTest {
 
-    private final Long PATIENT_1_SSN = 1L;
-    private final Long PATIENT_2_SSN = 2L;
-    private final Long PATIENT_3_SSN = 3L;
+    private final Integer PATIENT_1_SSN = 1;
+    private final Integer PATIENT_2_SSN = 2;
+    private final Integer PATIENT_3_SSN = 3;
 
-    private final HashMap<Long, ICUPatient> icuPatients = new HashMap<>();
+    private final HashMap<Integer, ICUPatient> icuPatients = new HashMap<>();
 
     @Test
     public void testCEPConfigThroughCode() {
@@ -76,7 +76,7 @@ public class LowOxygenBloodLevelTest {
 
     private void lowOxygenLevelTest(KieSession kieSession) {
 
-        final HashMap<Long, FactHandle> facts = fillTheMapInsertInSession(kieSession);
+        final HashMap<Integer, FactHandle> facts = fillTheMapInsertInSession(kieSession);
         final SessionPseudoClock clock = kieSession.getSessionClock();
 
         for (int index = 0; index < 30; index++) {
@@ -114,27 +114,27 @@ public class LowOxygenBloodLevelTest {
         assertThat(numOfLowOxygenLevelEvent.size(), equalTo(1));
     }
 
-    private HashMap<Long, FactHandle> fillTheMapInsertInSession(KieSession kieSession) {
+    private HashMap<Integer, FactHandle> fillTheMapInsertInSession(KieSession kieSession) {
         icuPatients.clear();
         icuPatients.put(PATIENT_1_SSN, createPatient(PATIENT_1_SSN));
         icuPatients.put(PATIENT_2_SSN, createPatient(PATIENT_2_SSN));
         icuPatients.put(PATIENT_3_SSN, createPatient(PATIENT_3_SSN));
 
-        final HashMap<Long, FactHandle> patientsFacts = new HashMap<>();
+        final HashMap<Integer, FactHandle> patientsFacts = new HashMap<>();
         for (ICUPatient patient : icuPatients.values()) {
             patientsFacts.put(patient.getSsn(), kieSession.insert(patient));
         }
         return patientsFacts;
     }
 
-    private ICUPatient createPatient(Long ssn) {
+    private ICUPatient createPatient(Integer ssn) {
         final ICUPatient patient = new ICUPatient();
         patient.setSsn(ssn);
         patient.setBloodOxygenLevel(70);
         return patient;
     }
 
-    private BloodOxygenLevelChangeEvent simulateBloodChange(Long ssn, double change) {
+    private BloodOxygenLevelChangeEvent simulateBloodChange(Integer ssn, double change) {
         icuPatients.get(ssn).setBloodOxygenLevel(icuPatients.get(ssn).getBloodOxygenLevel() + change);
         return new BloodOxygenLevelChangeEvent(ssn, change);
     }
